Split up strategies to try to clean up the code.
