+++ src/transformd/__init__.py
@@ -1,4 +1,5 @@
-from transformd.transformer import InvalidSpecError, Transformer
+from transformd.exceptions import InvalidSpecError
+from transformd.transformer import Transformer
 
 __all__ = [
     "Transformer",
