+++ src/transformd/exceptions.py
@@ -1,0 +1,2 @@
+class InvalidSpecError(Exception):
+    pass
